backend/foodgram/requirements.txt
asgiref==3.7.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==2.0.12
cryptography==42.0.5
defusedxml==0.8.0rc2
Django==3.2.16
django-colorfield==0.11.0
django-cors-headers==3.13.0
django-filter==23.4
django-templated-mail==1.1.1
djangorestframework==3.12.4
djoser==2.2.2
idna==3.6
oauthlib==3.2.2
pillow==10.2.0
psycopg2-binary==2.9.9
pycparser==2.21
PyJWT==2.7.0
python3-openid==3.2.0
pytz==2024.1
requests==2.26.0
requests-oauthlib==2.0.0
social-auth-app-django==5.4.0
social-auth-core==4.5.3
sqlparse==0.4.4
typing_extensions==4.9.0
urllib3==1.26.18
